Simplify test-cases, by using mp_set_l. Also fix use of types in 2 other files

+++ etc/mersenne.c
@@ -123,7 +123,7 @@ int main(void)
          tt = clock() - tt;
 
          /* display if prime */
-         printf("2^%-5ld - 1 is prime, test took %ld ticks\n", k, tt);
+         printf("2^%-5ld - 1 is prime, test took %ld ticks\n", k, (long)tt);
       }
 
       /* goto next odd exponent */
